Renaming local module.
