+++ android/widgets/build.gradle.kts
@@ -45,7 +45,7 @@ dependencies {
 
     implementation(project(":core:displaymodels"))
     implementation(project(":core:models"))
-    implementation(project(":data:local"))
+    implementation(project(":data:local-sqldelight"))
     implementation(project(":data:match"))
     implementation(project(":data:octanegg"))
     implementation(project(":data:remote"))
