+++ data/event/src/commonTest/kotlin/com/adammcneilly/pocketleague/data/event/SQLDelightEventServiceTest.kt
@@ -1,0 +1,8 @@
+package com.adammcneilly.pocketleague.data.event
+
+class SQLDelightEventServiceTest {
+
+    @Test
+    fun getUpcmingEvents() {
+    }
+}
