+++ data/local-sqldelight/src/androidMain/AndroidManifest.xml
