+++ data/local-sqldelight/src/commonMain/kotlin/com/adammcneilly/pocketleague/data/local/sqldelight/DatabaseDriverFactory.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.pocketleague.data.local
+package com.adammcneilly.pocketleague.data.local.sqldelight
 
 import com.squareup.sqldelight.db.SqlDriver
