+++ data/local-sqldelight/src/commonMain/kotlin/com/adammcneilly/pocketleague/data/local/sqldelight/mappers/LocalMatchMappers.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.pocketleague.data.local.mappers
+package com.adammcneilly.pocketleague.data.local.sqldelight.mappers
 
 import com.adammcneilly.pocketleague.core.models.Match
 import com.adammcneilly.pocketleague.sqldelight.LocalMatch
