+++ data/local-sqldelight/src/commonMain/sqldelight/com/adammcneilly/pocketleague/sqldelight/LocalTeam.sq
