+++ data/team/build.gradle.kts
@@ -10,7 +10,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(project(":core:models"))
-                implementation(project(":data:local"))
+                implementation(project(":data:local-sqldelight"))
                 implementation(project(":data:octanegg"))
                 implementation(project(":data:remote"))
                 implementation(libs.kotlinx.coroutines.core)
