+++ settings.gradle.kts
@@ -15,7 +15,7 @@ include(":core:models-test")
 include(":data:event")
 include(":data:event-test")
 include(":data:game")
-include(":data:local")
+include(":data:local-sqldelight")
 include(":data:match")
 include(":data:match-test")
 include(":data:octanegg")
