+++ shared/build.gradle.kts
@@ -15,7 +15,7 @@ kotlin {
                 implementation(project(":core:models"))
                 implementation(project(":data:event"))
                 implementation(project(":data:game"))
-                implementation(project(":data:local"))
+                implementation(project(":data:local-sqldelight"))
                 implementation(project(":data:match"))
                 implementation(project(":data:octanegg"))
                 implementation(project(":data:remote"))
