+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/di/DataModule.kt
@@ -2,7 +2,7 @@ package com.adammcneilly.pocketleague.shared.di
 
 import com.adammcneilly.pocketleague.data.event.EventRepository
 import com.adammcneilly.pocketleague.data.game.GameService
-import com.adammcneilly.pocketleague.data.local.PocketLeagueDB
+import com.adammcneilly.pocketleague.data.local.sqldelight.PocketLeagueDB
 import com.adammcneilly.pocketleague.data.match.MatchService
 import com.adammcneilly.pocketleague.data.team.TeamRepository
