+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/di/ProdDataModule.kt
@@ -6,8 +6,8 @@ import com.adammcneilly.pocketleague.data.event.OfflineFirstEventRepository
 import com.adammcneilly.pocketleague.data.event.SQLDelightEventService
 import com.adammcneilly.pocketleague.data.game.GameService
 import com.adammcneilly.pocketleague.data.game.OctaneGGGameService
-import com.adammcneilly.pocketleague.data.local.DatabaseDriverFactory
-import com.adammcneilly.pocketleague.data.local.PocketLeagueDB
+import com.adammcneilly.pocketleague.data.local.sqldelight.DatabaseDriverFactory
+import com.adammcneilly.pocketleague.data.local.sqldelight.PocketLeagueDB
 import com.adammcneilly.pocketleague.data.match.MatchService
 import com.adammcneilly.pocketleague.data.match.OfflineFirstMatchService
 import com.adammcneilly.pocketleague.data.octanegg.OctaneGGAPIClient
